+++ crates/zng-env/src/lib.rs
@@ -26,30 +26,49 @@ lazy_static! {
     static ref ABOUT: About = About::fallback_name();
 }
 
-/// Init [`about`] from cargo manifest and handles witch process to execute.
+/// Inits process metadata, calls process start handlers and defines the process lifetime in `main`.
 ///
-/// See [`About`] docs for what Cargo.toml values are read.
+/// This **must** be called in main.
 ///
-/// # Processes
+/// Init [`about`] an [`About`] for the process metadata. See [`on_process_start!`] for process start handlers.
+/// See [`on_process_exit`] for exit handlers called at the end of the `main` function.
+///
+/// # Process Start
 ///
 /// A single Zng executable can be built with multiple components that spawn different instances
 /// of the executable that must run as different processes. If the current instance is requested
 /// by component `init!` runs it and exits the process, never returning flow to the normal main function.
+///
+/// ```
+/// # mod zng { pub mod env { pub use zng_env::*; } }
+/// fn main() {
+///     println!("print in all processes");
+///     zng::env::init!();
+///     println!("print only in the app-process");
+///
+///     // directories are available after `init!`.
+///     let _res = zng::env::res("");
+///     
+///     // APP.defaults().run(...);
+///
+///     // on_exit handlers are called here
+/// }
+/// ```
 #[macro_export]
 macro_rules! init {
     () => {
-        $crate::init_parse!($crate)
+        let _on_main_exit = $crate::init_parse!($crate);
     };
 }
 #[doc(hidden)]
 pub use zng_env_proc_macros::init_parse;
 
 #[doc(hidden)]
-pub fn init(about: About) {
+pub fn init(about: About) -> impl Drop {
     if lazy_static_init(&ABOUT, about).is_err() {
         panic!("env already inited, env::init must be the first call in the process")
     }
-    process_init();
+    process_init()
 }
 
 /// Metadata about the app and main crate.