+++ crates/zng-view/src/lib.rs
@@ -151,22 +151,21 @@ use zng_view_api::{
 
 use rustc_hash::FxHashMap;
 
-/// Name of the view-process in [`zng_env::process_main!`].
 #[cfg(feature = "ipc")]
-pub const ZNG_ENV_VIEW_PROCESS: &str = "zng-view/view-process";
-
-#[cfg(feature = "ipc")]
-zng_env::process_main!(ZNG_ENV_VIEW_PROCESS => view_process_main);
+zng_env::on_process_start!(|_| view_process_main());
 
 /// Runs the view-process server.
 ///
 /// Note that this only needs to be called if the view-process is not built on the same executable, if
 /// it is you only need to call [`zng_env::init!`] at the beginning of the executable main.
 #[cfg(feature = "ipc")]
-pub fn view_process_main() -> ! {
-    std::panic::set_hook(Box::new(init_abort));
+pub fn view_process_main() {
+    let config = match ViewConfig::from_env() {
+        Some(c) => c,
+        None => return,
+    };
 
-    let config = ViewConfig::from_env().expect("view-process missing env config");
+    std::panic::set_hook(Box::new(init_abort));
     config.assert_version(false);
     let c = ipc::connect_view_process(config.server_name).expect("failed to connect to app-process");
 
@@ -181,13 +180,13 @@ pub fn view_process_main() -> ! {
         App::run_headed(c, ext);
     }
 
-    std::process::exit(0)
+    zng_env::exit(0)
 }
 
 #[cfg(feature = "ipc")]
 #[doc(hidden)]
 #[no_mangle]
-pub extern "C" fn extern_init() {
+pub extern "C" fn extern_view_process_main() {
     std::panic::set_hook(Box::new(ffi_abort));
     view_process_main()
 }
@@ -226,7 +225,7 @@ pub fn run_same_process_extended(run_app: impl FnOnce() + Send + 'static, ext: f
                     // This workaround ensures that don't become a zombie process.
                     thread::sleep(std::time::Duration::from_secs(5));
                     eprintln!("run_same_process did not exit after 5s of a fatal panic, exiting now");
-                    std::process::exit(101);
+                    zng_env::exit(101);
                 });
                 // Propagate panic in case the normal disconnect/shutdown handler works.
                 std::panic::resume_unwind(e);
@@ -275,7 +274,7 @@ fn panic_hook(info: &std::panic::PanicInfo, details: &str) {
         crate::util::set_suppressed_panic(panic);
     } else {
         eprintln!("{panic}\n{details}");
-        std::process::exit(101) // Rust panic exit code.
+        zng_env::exit(101) // Rust panic exit code.
     }
 }